requirements.txt
fastapi
uvicorn
sqlmodel
python-jose[cryptography]
passlib
python-multipart
